Subscription: added autherized users based subscirption storing

+++ backend/src/main/java/com/goodkiritsu/backend/subscription/SubscriptionService.java
@@ -1,5 +1,9 @@
 package com.goodkiritsu.backend.subscription;
 
+import com.goodkiritsu.backend.Security.User;
+import org.jspecify.annotations.Nullable;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -14,15 +18,23 @@ public class SubscriptionService {
     }
 
     public List<Subscription> getAllSubscriptions() {
-        List<Subscription> subs = subscriptionRepo.findAll();
+
+        List<Subscription> subs = subscriptionRepo.findByUser(getCurrentUser());
         return subs;
     }
 
     public void addSubscription(Subscription sub) {
+        sub.setUser(getCurrentUser());
         subscriptionRepo.save(sub);
     }
 
     public void deleteSubscription(Long id) {
+
+        Subscription sub = subscriptionRepo.findById(id)
+                        .orElseThrow(() -> new RuntimeException("Subscription not found"));
+        if (!sub.getUser().getId().equals(getCurrentUser().getId())) {
+            throw new RuntimeException("Not authorized to delete this subscription");
+        }
         subscriptionRepo.deleteById(id);
     }
 
@@ -34,4 +46,10 @@ public class SubscriptionService {
         }
         return sum;
     }
+
+    private User getCurrentUser() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        return (User) authentication.getPrincipal();
+    }
+
 }
